create module 'split_queryset'
