+++ your_finance_helper/main_page/views.py
@@ -6,35 +6,15 @@ from django.urls import reverse_lazy
 from .forms import AddIncomeForm, AddOutcomeForm, AddNewSectionForm, AddNewCategoryForm, AddNewNameForm
 from reports.forms import DateWidgetForm
 from datetime import datetime
+from .split_queryset import split_queryset
 
 
 class IndexView(View):
     def get(self, request):
-        def split_queryset(data):
-            sum_all = 0
-            dict_section = {}
-            dict_category = {}
-            dict_name = {}
-            for i in data:
-                dict_section[i['id_section__section']] = {
-                    'id_section__id': i['id_section__id']
-                }
-                dict_category[i['id_category__category']] = {
-                    'id_category__id': i['id_category__id'],
-                    'id_category__to_section': i['id_category__to_section']
-                }
-                dict_name[i['id_name__name']] = {
-                    'id_name__to_category': i['id_name__to_category'],
-                    'sum': i['sum']
-                }
-                sum_all += i['sum']
-            return dict_section, dict_category, dict_name, sum_all
-
         in_dict_section, in_dict_category, in_dict_name, in_sum_all = split_queryset(
             GeneralTable.objects.filter(
                 date__range=[datetime.now().replace(day=1).date(),
-                             datetime.now().date()]
-            ).filter(
+                             datetime.now().date()]).filter(
                 type_of_transaction='IN').values(
                     'id_section__id',
                     'id_section__section',
@@ -106,26 +86,6 @@ class AddNewNameView(CreateView):
 
 class AnotherMainPeriod(View):
     def get(self, request):
-        def split_queryset(data):
-            sum_all = 0
-            dict_section = {}
-            dict_category = {}
-            dict_name = {}
-            for i in data:
-                dict_section[i['id_section__section']] = {
-                    'id_section__id': i['id_section__id']
-                }
-                dict_category[i['id_category__category']] = {
-                    'id_category__id': i['id_category__id'],
-                    'id_category__to_section': i['id_category__to_section']
-                }
-                dict_name[i['id_name__name']] = {
-                    'id_name__to_category': i['id_name__to_category'],
-                    'sum': i['sum']
-                }
-                sum_all += i['sum']
-            return dict_section, dict_category, dict_name, sum_all
-
         if request.method == 'GET':
             form = DateWidgetForm(request.GET)
             if form.is_valid():
